+++ src/main/java/com/umc/mada/custom/dto/ItemsElementResponse.java
@@ -5,14 +5,14 @@ import lombok.*;
 
 @Getter
 public class ItemsElementResponse {
-    private Long id;
+    private int id;
     private String itemType;
     private String itemUnlockCondition; //TODO: 미션이라면 미션 내용도 반환하기 위해 mission변수 추가해야함
     private String filePath;
     private boolean have; //유저가 해당 아이템을 소유하고 있는지 여부
 
     @Builder
-    private ItemsElementResponse(Long id, String itemType, String itemUnlockCondition,String filePath, boolean have){
+    private ItemsElementResponse(int id, String itemType, String itemUnlockCondition,String filePath, boolean have){
         this.id = id;
         this.itemType = itemType;
         this.itemUnlockCondition = itemUnlockCondition;
